requirements.txt
# Core
anthropic>=0.40.0
python-dotenv>=1.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
rich>=13.0.0

# Search & scrape
tavily-python>=0.3.0
firecrawl-py>=1.0.0
